Patch issues with contrib/unzip building
This patch addresses two outstanding issues with using the files
in contrib/unzip.

1) When unzip.h is used with Qt, it needs to use the same version of
zlib that Qt uses, but that also means that it needs to import the
header outside of the extern "C" block. Previously we had another
copy of this header outside of the assimp repository code, but this
file has gotten out of sync with what is in the upstream repository.

2) ioapi.h previous has an issue where it's not possible to build for
Android when targeting __ANDROID_API__ < 24. So we patch this file to
account for this. There is an outstanding issue upstream #5408 that
when fixed should make this patch unnecessary.

Change-Id: I4c0c11d198c91e267af1f2310f3abfa403d2fce4
Reviewed-by: Andy Nichols <andy.nichols@qt.io>